+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- sketchjob CLI options for selecting wrapper script
+- Templates for both job command and wrapper script creation
 - ``jobarchitect.backends.render_script`` function
 - Hosting of docs on `readthedocs <http://jobarchitect.readthedocs.io/>`_
 - API documentation to sphinx generated docs